EmailValidatorTests added
